pyproject.toml
[tool.poetry]
name = "pdfbot"
version = "0.1.0"
description = ""
authors = ["Justin Levi Winter <justinlevi@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^1.0.0"
streamlit = "^1.22.0"
watchdog = "^3.0.0"
pypdf2 = "^3.0.1"
langchain = "^0.0.173"
openai = "^0.27.6"
tiktoken = "^0.4.0"
faiss-cpu = "^1.7.4"

[tool.poetry.group.dev.dependencies]
pytest = "^7.3.1"
mock = "^5.0.2"

[dev-dependencies]
streamlit = "*"
PyPDF2 = "*"
langchain = "*"
tiktoken = "*"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
